Removed the need for SelectItem and made the Bootstrap dependency optional
SelectItem has now been completely removed.
The only time any objects are now created are when the incoming array
consists of strings, or if it's multidimensional, otherwise the array is
unchanged.

Also improved the styling a bit and made the Bootstrap dependency
optional by providing input parameters for changing the defaults to
whatever wanted.
The ng2-select specific css classes are still there though, as they are
needed.